+++ include/phoxi_camera/CoordinatesSpaceTypes.h
@@ -1,0 +1,13 @@
+//
+// Created by controller on 1/12/18.
+//
+
+#ifndef PROJECT_COORDINATESSPACE_H
+#define PROJECT_COORDINATESSPACE_H
+
+#include <PhoXi.h>
+namespace phoxi_camera{
+    typedef pho::api::PhoXiCoordinateSpace CoordinatesSpace;
+}
+
+#endif //PROJECT_COORDINATESSPACE_H