+++ CHANGELOG.rst
@@ -14,6 +14,9 @@ Changelog for package phoxi_camera
 * GetHardwareIdentification.srv added message success response parameter.
 * GetSuportedCapturingModes.srv added message success response parameter.
 * Added V2/is_connected, V2/is_acquiring, V2/start_acquisition, V2/stop_acquisition services.
+* trigger_image service publish frame to Ros topics automatically
+* get_frame service will take new frane on negative number input and publish it, on invalid frame number success will be false and message filled with error
+* V2/set_transformation and V2/set_coordination_space services added
 
 1.1.4 (2016-10-24)
 ------------------